IdP: remove dummy placeholder logo from logout.jsp
We don't have any logo installed in the IdP webapp and it would be extra
maintenance adding it to install instructions.  As this page doesn't
have other branding on it, not worth throwing in just the logo.

And pulling the logo from the VHO app would be too brittle - would
break whenever the applicationbase plugin changes - the URL is
/plugins/aaf-application-base-0.11.0/images/logo.jpg

The instructions for the dummy placeholder logo say "replace or remove"
- so removing.