IdP: remove VHR session cookie from logout.jsp
Make logout.jsp remove the session cookie set by the VHR app for
VhrFilter running inside the IdP app.

This cookie automatically expires after 2 minutes, but let's remove it
explicitly.

Note that for SOAP SLO, this page would not be invoked and the cookie
would not be removed ... but then, it expires after 2 minutes.