LoginApi: include authnInstant in JSON response
Modify LoginApiController to include authentication instant in the JSON
response to the confirmsession API request.

The instant is passed as a string formatted in ISO8601 with seconds
included to three decimal places (milliseconds).

Default JSON serialization would skip the second fractions, so create a
custom SimpleDateFormat instance.  As we also need to set the timezone
on that instance to UTC, define a simple method that creates the
instance and is called upon the LoginApiController initialization.

To record the authentication instant, modify how sessions are stored by
LoginService in loginCache: instead of storing just a String username
value, define a new class AuthnTuple (under src/groovy) to hold the
username and the Date authentication instant.

Modify the loginService.sessionRemoteUser method to return the tuple instead.

The authentication instant is recorded and stored in loginService.establishSession.

Overall, this change supports the handling of forceAuthn introduced in 7d58c7a.