Login: accept and display servicename of the SP
Accept servicename (and relyingparty) parameters sent from the IdP
integration module.  (The relyingparty parameter had been sent for long;
the servicename parameter has been added in 6963eab).

Store the parameters in the http session.

On the login form, render these parameters if they are present (and if
the form is being accessed from the login controller - to avoid dropping
the service name onto the MyAccount login form).

When rendering the service name, apply the "servicename" CSS class to
the name and the login prompt - and from CSS, apply "text-align:
center;".