LoginController: add logout method
This method zaps the L1 and L2 sessions - both internally (in
TwoStepSession and LoginCache) as well as zapping the respective cookie.

This can be invoked at

https://virtualhome.server/login/logout?slourl=/target

and the intended use is to hook into the IdP Logout flow with a redirect
pointing back to the flow.

The following code in
/opt/shibboleth-idp/system/flows/logout/logout-flow.xml does the job:

    <view-state id="AppLogoutView" view="externalRedirect:serverRelative:#{ '/login/logout?slourl='+new org.apache.commons.codec.net.URLCodec().encode(flowExecutionUrl+ '&amp;_eventId_proceed=1') }" >
        <transition on="proceed" to="LogoutView" />
    </view-state>

(and changing the target in NextRelyingPartyContext from LogoutView to AppLogoutView)