new: ShibInt3: add support for MFA signalling
Support signalling whether MFA was used through the set of principals returned.

This mode gets activated by setting the `mfaPrincipalName` initialisation parameter for the `VhrRemoteUserAuthServlet`.

In this mode, the servlet iterates over the principals (authentication contexts)
configured as supported by the authentication flow, and adds them to the returned
Subject according to the following rules:
 * the principal matching the mfaPrincipalName parameter is only added when MFA is used by the user account.
 * all other principals are assumed to represent Single Authentication Factor (SFA) flows and are added regardless whether MFA was used in the session.

Note that this mode requires version of the VHO app extended to include the MFA status in the authentication response.

Note also that it requires the IdP to have the RemoteUser authentication flow
configured with correct set of supportedPrincipals, extended to include the
principal referred to by `mfaPrincipalName`
(recommended to use `https://refeds.org/profile/mfa`).

It also requires the `RemoteUser` authentication flow configured not to automatically
add all support principals (which would conflict with the approach of adding
the MFA principal only when MFA is actually used).
This is achieved by setting the `addDefaultPrincipals` bean (or property) to False.